Remove BootstrapVueProcessor class
- No need to capture any state, having the transformation functions will do
